Factor out MediaTrack
This should make way for DataTrack.
