+++ lib/cancelableroompromise.js
@@ -4,7 +4,7 @@ var CancelablePromise = require('./util/cancelablepromise');
 
 /**
  * Create a {@link CancelablePromise<Room>}.
- * @param {function(function(Array<LocalTracks>): CancelablePromise<RoomSignaling>):
+ * @param {function(function(Array<LocalTrack>): CancelablePromise<RoomSignaling>):
  *   Promise<function(): CancelablePromise<RoomSignaling>>} getLocalTracks
  * @param {function(Array<LocalTrack>): LocalParticipant} createLocalParticipant
  * @param {function(Array<LocalTrack>): CancelablePromise<RoomSignaling>} createRoomSignaling
