+++ lib/connect.js
@@ -29,13 +29,14 @@ var didPrintSafariWarning = false;
  *   <br><br>
  *   By default, this will automatically acquire an array containing a
  *   {@link LocalAudioTrack} and {@link LocalVideoTrack} before connecting to
- *   the {@link Room}. The {@link LocalTrack}s will be stopped when you
- *   disconnect from the {@link Room}.
+ *   the {@link Room}. These will be stopped when you disconnect from the
+ *   {@link Room}.
  *   <br><br>
  *   You can override the default behavior by specifying
- *   <code>options</code>. For example, rather than acquiring {@link LocalTrack}s
- *   automatically, you can pass your own array which you can stop yourself.
- *   See {@link ConnectOptions} for more information.
+ *   <code>options</code>. For example, rather than acquiring a
+ *   {@link LocalAudioTrack} and {@link LocalVideoTrack} automatically, you can
+ *   pass your own array which you can stop yourself. See {@link ConnectOptions}
+ *   for more information.
  * @param {string} token - The Access Token string
  * @param {ConnectOptions} [options] - Options to override the default behavior
  * @returns {CancelablePromise<Room>}
