+++ lib/createlocaltrack.js
@@ -4,10 +4,10 @@ var defaultCreateLocalTracks = require('./createlocaltracks');
 var DEFAULT_LOG_LEVEL = require('./util/constants').DEFAULT_LOG_LEVEL;
 
 /**
- * Request a {@link LocalTrack} of a particular kind.
- * @param {string} kind - 'audio' or 'video'
+ * Request a {@link LocalAudioTrack} or {@link LocalVideoTrack}.
+ * @param {Track.Kind} kind - "audio" or "video"
  * @param {CreateLocalTrackOptions} [options]
- * @returns {Promise<LocalTrack>}
+ * @returns {Promise<LocalAudioTrack|LocalVideoTrack>}
  * @private
  */
 function createLocalTrack(kind, options) {
