+++ lib/localparticipant.js
@@ -540,6 +540,7 @@ LocalParticipant.prototype.unpublishTracks = function unpublishTracks(tracks) {
  */
 
 /**
+ * @private
  * @param {Map<Track.SID, TrackPublication>} trackPublications
  * @param {Track|MediaStreamTrack} track
  * @returns {TrackPublication?} trackPublication
