+++ lib/media/track/audiotrack.js
@@ -1,28 +1,25 @@
 'use strict';
 
 var inherits = require('util').inherits;
-var Track = require('./');
+var MediaTrack = require('./mediatrack');
 
 /**
  * Construct an {@link AudioTrack} from a MediaStreamTrack.
  * @class
- * @classdesc An {@link AudioTrack} is a {@link Track} representing audio.
- * @extends Track
+ * @classdesc An {@link AudioTrack} is a {@link MediaTrack} representing audio.
+ * @extends MediaTrack
  * @param {MediaStreamTrack} mediaStreamTrack
  * @param {{log: Log}} options
+ * @property {Track.Kind} kind - "audio"
  */
 function AudioTrack(mediaStreamTrack, options) {
-  Track.call(this, mediaStreamTrack, options);
+  MediaTrack.call(this, mediaStreamTrack, options);
 }
 
-inherits(AudioTrack, Track);
+inherits(AudioTrack, MediaTrack);
 
-AudioTrack.prototype.toString = function toString() {
-  return '[AudioTrack #' + this._instanceId + ': ' + this.id + ']';
-};
+AudioTrack.prototype.attach = MediaTrack.prototype.attach;
 
-AudioTrack.prototype.attach = Track.prototype.attach;
-
-AudioTrack.prototype.detach = Track.prototype.detach;
+AudioTrack.prototype.detach = MediaTrack.prototype.detach;
 
 module.exports = AudioTrack;
