+++ lib/media/track/localaudiotrack.js
@@ -2,28 +2,28 @@
 
 var AudioTrack = require('./audiotrack');
 var inherits = require('util').inherits;
-var LocalTrack = require('./localtrack');
+var LocalMediaTrack = require('./localmediatrack');
 
 /**
  * Construct a {@link LocalAudioTrack} from a MediaStreamTrack.
  * @class
  * @classdesc A {@link LocalAudioTrack} is an {@link AudioTrack} representing
- * audio that your {@link LocalParticipant} sends to a {@link Room}.
+ *   audio that your {@link LocalParticipant} sends to a {@link Room}.
  * @extends {AudioTrack}
- * @extends {LocalTrack}
+ * @extends {LocalMediaTrack}
  * @param {MediaStreamTrack} mediaStreamTrack - The underlying MediaStreamTrack
- * @param {LocalTrackOptions} options - {@link LocalTrack} options
+ * @param {LocalTrackOptions} [options] - {@link LocalTrack} options
  */
 function LocalAudioTrack(mediaStreamTrack, options) {
   if (!(this instanceof LocalAudioTrack)) {
     return new LocalAudioTrack(mediaStreamTrack, options);
   }
-  LocalTrack.call(this, AudioTrack, mediaStreamTrack, options);
+  LocalMediaTrack.call(this, AudioTrack, mediaStreamTrack, options);
 }
 
 inherits(LocalAudioTrack, AudioTrack);
 
-LocalAudioTrack.prototype._end = LocalTrack.prototype._end;
+LocalAudioTrack.prototype._end = LocalMediaTrack.prototype._end;
 
 LocalAudioTrack.prototype.toString = function toString() {
   return '[LocalAudioTrack #' + this._instanceId + ': ' + this.id + ']';
@@ -39,26 +39,27 @@ LocalAudioTrack.prototype.attach = function attach(el) {
  * Disable the {@link LocalAudioTrack}. This is effectively "mute".
  * @method
  * @returns {this}
- * @fires Track#disabled
+ * @fires MediaTrack#disabled
  */
-LocalAudioTrack.prototype.disable = LocalTrack.prototype.disable;
+LocalAudioTrack.prototype.disable = LocalMediaTrack.prototype.disable;
 
 /**
  * Enable the {@link LocalAudioTrack}. This is effectively "unmute".
  * @method
  * @returns {this}
- * @fires Track#enabled
+ * @fires MediaTrack#enabled
 *//**
- * Enable or disable the {@link LocalAudioTrack}. This is effectively "unmute" or
- * "mute".
+ * Enable or disable the {@link LocalAudioTrack}. This is effectively "unmute"
+ * or "mute".
  * @method
- * @param {boolean} [enabled] - Specify false to mute the {@link LocalAudioTrack}
+ * @param {boolean} [enabled] - Specify false to mute the
+ *   {@link LocalAudioTrack}
  * @returns {this}
- * @fires Track#disabled
- * @fires Track#enabled
+ * @fires MediaTrack#disabled
+ * @fires MediaTrack#enabled
  */
-LocalAudioTrack.prototype.enable = LocalTrack.prototype.enable;
+LocalAudioTrack.prototype.enable = LocalMediaTrack.prototype.enable;
 
-LocalAudioTrack.prototype.stop = LocalTrack.prototype.stop;
+LocalAudioTrack.prototype.stop = LocalMediaTrack.prototype.stop;
 
 module.exports = LocalAudioTrack;
