+++ lib/media/track/localaudiotrackpublication.js
@@ -14,6 +14,7 @@ var LocalTrackPublication = require('./localtrackpublication');
  * @param {function(LocalTrackPublication): void} unpublish - The callback
  *    that unpublishes the {@link LocalTrackPublication}
  * @param {LocalTrackPublicationOptions} options - {@link LocalTrackPublication} options
+ * @property {Track.Kind} kind - "audio"
  * @property {LocalAudioTrack} track - the {@link LocalAudioTrack}
  */
 function LocalAudioTrackPublication(sid, track, unpublish, options) {
