+++ lib/media/track/localtrackpublication.js
@@ -8,16 +8,17 @@ var nInstances = 0;
 /**
  * Construct a {@link LocalTrackPublication}.
  * @class
- * @classdesc A {@link LocalTrackPublication} is a {@link LocalTrack} that has been
- * published to a {@link Room}.
+ * @classdesc A {@link LocalTrackPublication} is a {@link LocalTrack} that
+ *   has been published to a {@link Room}.
  * @param {Track.SID} sid - SID assigned to the published {@link LocalTrack}
  * @param {LocalTrack} track - the {@Link LocalTrack}
  * @param {function(LocalTrackPublication): void} unpublish - The callback
- *    that unpublishes the {@link LocalTrackPublication}
- * @param {LocalTrackPublicationOptions} options - {@link LocalTrackPublication} options
- * @property {string} kind - kind of the published {@link LocalTrack}; one of
- *   'audio' or 'video'
- * @property {Track.SID} sid - SID assigned to the published {@link LocalTrack}
+ *   that unpublishes the {@link LocalTrackPublication}
+ * @param {LocalTrackPublicationOptions} options - {@link LocalTrackPublication}
+ *   options
+ * @property {Track.Kind} kind - kind of the published {@link LocalTrack}
+ * @property {Track.SID} sid - SID assigned to the published
+ *   {@link LocalTrack}
  * @property {LocalTrack} track - the {@link LocalTrack}
  */
 function LocalTrackPublication(sid, track, unpublish, options) {
@@ -58,8 +59,8 @@ LocalTrackPublication.prototype.toString = function toString() {
 
 /**
  * Unpublish a {@link LocalTrackPublication}. This means that the media
- * from this {@link LocalTrackPublication} is no longer available to the {@link Room}'s
- * {@link RemoteParticipant}s.
+ * from this {@link LocalTrackPublication} is no longer available to the
+ * {@link Room}'s {@link RemoteParticipant}s.
  * @returns {this}
  */
 LocalTrackPublication.prototype.unpublish = function unpublish() {
