+++ lib/media/track/localvideotrack.js
@@ -1,29 +1,29 @@
 'use strict';
 
 var inherits = require('util').inherits;
-var LocalTrack = require('./localtrack');
+var LocalMediaTrack = require('./localmediatrack');
 var VideoTrack = require('./videotrack');
 
 /**
  * Construct a {@link LocalVideoTrack} from MediaStreamTrack.
  * @class
  * @classdesc A {@link LocalVideoTrack} is a {@link VideoTrack} representing
- * audio that your {@link LocalParticipant} sends to a {@link Room}.
+ *   video that your {@link LocalParticipant} sends to a {@link Room}.
  * @extends {VideoTrack}
- * @extends {LocalTrack}
+ * @extends {LocalMediaTrack}
  * @param {MediaStreamTrack} mediaStreamTrack - The underlying MediaStreamTrack
- * @param {LocalTrackOptions} options - {@link LocalTrack} options
+ * @param {LocalTrackOptions} [options] - {@link LocalTrack} options
  */
 function LocalVideoTrack(mediaStreamTrack, options) {
   if (!(this instanceof LocalVideoTrack)) {
     return new LocalVideoTrack(mediaStreamTrack, options);
   }
-  LocalTrack.call(this, VideoTrack, mediaStreamTrack, options);
+  LocalMediaTrack.call(this, VideoTrack, mediaStreamTrack, options);
 }
 
 inherits(LocalVideoTrack, VideoTrack);
 
-LocalVideoTrack.prototype._end = LocalTrack.prototype._end;
+LocalVideoTrack.prototype._end = LocalMediaTrack.prototype._end;
 
 LocalVideoTrack.prototype.toString = function toString() {
   return '[LocalVideoTrack #' + this._instanceId + ': ' + this.id + ']';
@@ -33,29 +33,30 @@ LocalVideoTrack.prototype.toString = function toString() {
  * Enable the {@link LocalVideoTrack}. This is effectively "unpause".
  * @method
  * @returns {this}
- * @fires Track#enabled
+ * @fires MediaTrack#enabled
 *//**
- * Enable or disable the {@link LocalVideoTrack}. This is effectively "unpause" or
- * "pause".
+ * Enable or disable the {@link LocalVideoTrack}. This is effectively "unpause"
+ * or "pause".
  * @method
- * @param {boolean} [enabled] - Specify false to pause the {@link LocalVideoTrack}
+ * @param {boolean} [enabled] - Specify false to pause the
+ *   {@link LocalVideoTrack}
  * @returns {this}
- * @fires Track#disabled
- * @fires Track#enabled
+ * @fires MediaTrack#disabled
+ * @fires MediaTrack#enabled
  */
-LocalVideoTrack.prototype.enable = LocalTrack.prototype.enable;
+LocalVideoTrack.prototype.enable = LocalMediaTrack.prototype.enable;
 
 /**
  * Disable the {@link LocalVideoTrack}. This is effectively "pause".
  * @method
  * @returns {this}
- * @fires Track#disabled
+ * @fires MediaTrack#disabled
  */
-LocalVideoTrack.prototype.disable = LocalTrack.prototype.disable;
+LocalVideoTrack.prototype.disable = LocalMediaTrack.prototype.disable;
 
 /**
- * See {@link LocalTrack#stop}.
+ * See {@link LocalMediaTrack#stop}.
  */
-LocalVideoTrack.prototype.stop = LocalTrack.prototype.stop;
+LocalVideoTrack.prototype.stop = LocalMediaTrack.prototype.stop;
 
 module.exports = LocalVideoTrack;
