+++ lib/media/track/localvideotrackpublication.js
@@ -14,6 +14,7 @@ var LocalTrackPublication = require('./localtrackpublication');
  * @param {function(LocalTrackPublication): void} unpublish - The callback
  *    that unpublishes the {@link LocalTrackPublication}
  * @param {LocalTrackPublicationOptions} options - {@link LocalTrackPublication} options
+ * @property {Track.Kind} kind - "video"
  * @property {LocalVideoTrack} track - the {@link LocalVideoTrack}
  */
 function LocalVideoTrackPublication(sid, track, unpublish, options) {
