+++ lib/media/track/remotevideotrack.js
@@ -1,16 +1,16 @@
 'use strict';
 
 var inherits = require('util').inherits;
-var RemoteTrack = require('./remotetrack');
+var RemoteMediaTrack = require('./remotemediatrack');
 var VideoTrack = require('./videotrack');
 
 /**
  * Construct an {@link RemoteVideoTrack} from a MediaStreamTrack.
  * @class
- * @classdesc A {@link RemoteVideoTrack} represents a media {@link VideoTrack} published
- * to the {@link Room} by a {@link RemoteParticipant}.
+ * @classdesc A {@link RemoteVideoTrack} represents a {@link VideoTrack}
+ *   published to the {@link Room} by a {@link RemoteParticipant}.
  * @extends VideoTrack
- * @extends RemoteTrack
+ * @extends RemoteMediaTrack
  * @param {MediaStreamTrack} mediaStreamTrack - The underlying MediaStreamTrack
  * @param {RemoteTrackSignaling} signaling - The {@link Track} signaling
  * @param {{log: Log}} options - The {@link RemoteTrack} options
@@ -19,7 +19,7 @@ function RemoteVideoTrack(mediaStreamTrack, signaling, options) {
   if (!(this instanceof RemoteVideoTrack)) {
     return new RemoteVideoTrack(mediaStreamTrack, signaling, options);
   }
-  RemoteTrack.call(this, VideoTrack, mediaStreamTrack, signaling, options);
+  RemoteMediaTrack.call(this, VideoTrack, mediaStreamTrack, signaling, options);
 }
 
 inherits(RemoteVideoTrack, VideoTrack);
@@ -28,6 +28,6 @@ RemoteVideoTrack.prototype.toString = function toString() {
   return '[RemoteVideoTrack #' + this._instanceId + ': ' + this.sid + ']';
 };
 
-RemoteVideoTrack.prototype._unsubscribe = RemoteTrack.prototype._unsubscribe;
+RemoteVideoTrack.prototype._unsubscribe = RemoteMediaTrack.prototype._unsubscribe;
 
 module.exports = RemoteVideoTrack;
