+++ lib/media/track/videotrack.js
@@ -1,20 +1,22 @@
 'use strict';
 
 var inherits = require('util').inherits;
-var Track = require('./');
+var MediaTrack = require('./mediatrack');
 
 /**
  * Construct a {@link VideoTrack} from MediaStreamTrack.
  * @class
- * @classdesc A {@link VideoTrack} is a {@link Track} representing video.
- * @extends Track
+ * @classdesc A {@link VideoTrack} is a {@link MediaTrack} representing video.
+ * @extends MediaTrack
  * @param {MediaStreamTrack} mediaStreamTrack
  * @param {{log: Log}} options
- * @property {VideoTrack#Dimensions} dimensions - The {@link VideoTrack}'s {@link VideoTrack#Dimensions}
+ * @property {VideoTrack#Dimensions} dimensions - The {@link VideoTrack}'s
+ *   {@link VideoTrack#Dimensions}
+ * @property {Track.Kind} kind - "video"
  * @fires VideoTrack#dimensionsChanged
  */
 function VideoTrack(mediaStreamTrack, options) {
-  Track.call(this, mediaStreamTrack, options);
+  MediaTrack.call(this, mediaStreamTrack, options);
   Object.defineProperties(this, {
     dimensions: {
       enumerable: true,
@@ -64,22 +66,18 @@ function dimensionsChanged(track, elem) {
     || track.dimensions.height !== elem.videoHeight;
 }
 
-inherits(VideoTrack, Track);
-
-VideoTrack.prototype.toString = function toString() {
-  return '[VideoTrack #' + this._instanceId + ': ' + this.id + ']';
-};
+inherits(VideoTrack, MediaTrack);
 
 VideoTrack.prototype._start = function _start(dummyEl) {
   this.dimensions.width = dummyEl.videoWidth;
   this.dimensions.height = dummyEl.videoHeight;
 
   this._log.debug('Dimensions:', this.dimensions);
-  return Track.prototype._start.call(this, dummyEl);
+  return MediaTrack.prototype._start.call(this, dummyEl);
 };
 
 VideoTrack.prototype.attach = function attach(el) {
-  el = Track.prototype.attach.call(this, el);
+  el = MediaTrack.prototype.attach.call(this, el);
   el.muted = true;
   return el;
 };
