+++ lib/util/index.js
@@ -41,6 +41,15 @@ function asLocalTrackPublication(track, sid, unpublish, options) {
 }
 
 /**
+ * Capitalize a word.
+ * @param {string} word
+ * @param {string} capitalized
+ */
+function capitalize(word) {
+  return word[0].toUpperCase() + word.slice(1);
+}
+
+/**
  * Finds the items in list1 that are not in list2.
  * @param {Array<*>|Map<*>|Set<*>} list1
  * @param {Array<*>|Map<*>|Set<*>} list2
@@ -372,23 +381,24 @@ function getSdpFormat() {
   return typeof mozRTCPeerConnection !== 'undefined' ? 'unified' : null;
 }
 
-module.exports.constants = constants;
-module.exports.asLocalTrack = asLocalTrack;
-module.exports.asLocalTrackPublication = asLocalTrackPublication;
-module.exports.difference = difference;
-module.exports.flatMap = flatMap;
-module.exports.guessBrowser = guessBrowser;
-module.exports.makeClientSIPURI = makeClientSIPURI;
-module.exports.makeServerSIPURI = makeServerSIPURI;
-module.exports.makeUUID = makeUUID;
-module.exports.oncePerTick = oncePerTick;
-module.exports.promiseFromEvents = promiseFromEvents;
-module.exports.getOrNull = getOrNull;
-module.exports.defer = defer;
-module.exports.delegateMethods = delegateMethods;
-module.exports.proxyProperties = proxyProperties;
-module.exports.legacyPromise = legacyPromise;
-module.exports.buildLogLevels = buildLogLevels;
-module.exports.trackClass = trackClass;
-module.exports.validateLocalTrack = validateLocalTrack;
-module.exports.getSdpFormat = getSdpFormat;
+exports.constants = constants;
+exports.asLocalTrack = asLocalTrack;
+exports.asLocalTrackPublication = asLocalTrackPublication;
+exports.capitalize = capitalize;
+exports.difference = difference;
+exports.flatMap = flatMap;
+exports.guessBrowser = guessBrowser;
+exports.makeClientSIPURI = makeClientSIPURI;
+exports.makeServerSIPURI = makeServerSIPURI;
+exports.makeUUID = makeUUID;
+exports.oncePerTick = oncePerTick;
+exports.promiseFromEvents = promiseFromEvents;
+exports.getOrNull = getOrNull;
+exports.defer = defer;
+exports.delegateMethods = delegateMethods;
+exports.proxyProperties = proxyProperties;
+exports.legacyPromise = legacyPromise;
+exports.buildLogLevels = buildLogLevels;
+exports.trackClass = trackClass;
+exports.validateLocalTrack = validateLocalTrack;
+exports.getSdpFormat = getSdpFormat;
