+++ scripts/docs.js
@@ -17,12 +17,15 @@ const publicClasses = [
   'lib/media/track/index.js',
   'lib/media/track/audiotrack.js',
   'lib/media/track/localaudiotrack.js',
-  'lib/media/track/localtrack.js',
+  'lib/media/track/localmediatrack.js',
   'lib/media/track/localvideotrack.js',
   'lib/media/track/localaudiotrackpublication.js',
+  'lib/media/track/localtrack.js',
   'lib/media/track/localtrackpublication.js',
   'lib/media/track/localvideotrackpublication.js',
+  'lib/media/track/mediatrack.js',
   'lib/media/track/remoteaudiotrack.js',
+  'lib/media/track/remotemediatrack.js',
   'lib/media/track/remotetrack.js',
   'lib/media/track/remotevideotrack.js',
   'lib/media/track/videotrack.js',
@@ -50,11 +53,12 @@ const privateConstructors = [
   'AudioTrack',
   'Room',
   'LocalParticipant',
-  'LocalTrack',
+  'LocalMediaTrack',
   'Participant',
   'LocalAudioTrackPublication',
   'LocalTrackPublication',
   'LocalVideoTrackPublication',
+  'MediaTrack',
   'RemoteAudioTrack',
   'RemoteTrack',
   'RemoteVideoTrack',
