+++ test/unit/spec/media/track/localmediatrack.js
@@ -3,7 +3,7 @@
 var assert = require('assert');
 var inherits = require('util').inherits;
 var EventEmitter = require('events').EventEmitter;
-var Track = require('../../../../../lib/media/track');
+var Track = require('../../../../../lib/media/track/mediatrack');
 var LocalAudioTrack = require('../../../../../lib/media/track/localaudiotrack');
 var LocalVideoTrack = require('../../../../../lib/media/track/localvideotrack');
 var sinon = require('sinon');
@@ -14,7 +14,7 @@ var log = require('../../../../lib/fakelog');
   ['LocalVideoTrack', LocalVideoTrack]
 ].forEach(pair => {
   var description = pair[0];
-  var LocalTrack = pair[1];
+  var LocalMediaTrack = pair[1];
   var kind = {
     LocalAudioTrack: 'audio',
     LocalVideoTrack: 'video'
@@ -34,20 +34,20 @@ var log = require('../../../../lib/fakelog');
         [
           [
             'when called without the "new" keyword',
-            () => LocalTrack(mediaStreamTrack)
+            () => LocalMediaTrack(mediaStreamTrack)
           ],
           [
             'when called with the "new" keyword',
-            () => new LocalTrack(mediaStreamTrack)
+            () => new LocalMediaTrack(mediaStreamTrack)
           ]
-        ].forEach(([ scenario, createLocalTrack ]) => {
+        ].forEach(([ scenario, createLocalMediaTrack ]) => {
           context(scenario, () => {
             it('should not throw', () => {
-              assert.doesNotThrow(createLocalTrack);
+              assert.doesNotThrow(createLocalMediaTrack);
             });
 
             it(`should return an instance of ${description}`, () => {
-              assert(createLocalTrack() instanceof LocalTrack);
+              assert(createLocalMediaTrack() instanceof LocalMediaTrack);
             });
           });
         });
@@ -56,7 +56,7 @@ var log = require('../../../../lib/fakelog');
 
     describe('.isEnabled', () => {
       it('should set the .isEnabled to the MediaStreamTrack\'s .enabled property', () => {
-        track = createTrack(LocalTrack, '1', kind[description]);
+        track = crateLocalMediaTrack(LocalMediaTrack, '1', kind[description]);
         assert.equal(track.isEnabled, track.mediaStreamTrack.enabled);
         track.mediaStreamTrack.enabled = !track.mediaStreamTrack.enabled;
         assert.equal(track.isEnabled, track.mediaStreamTrack.enabled);
@@ -65,7 +65,7 @@ var log = require('../../../../lib/fakelog');
 
     describe('.isStopped', () => {
       it('should set .isStopped based on the state of the MediaStreamTrack\'s .readyState property', () => {
-        track = createTrack(LocalTrack, '1', kind[description]);
+        track = crateLocalMediaTrack(LocalMediaTrack, '1', kind[description]);
         track.mediaStreamTrack.readyState = 'ended';
         assert(track.isStopped);
         track.mediaStreamTrack.readyState = 'live';
@@ -75,8 +75,8 @@ var log = require('../../../../lib/fakelog');
 
     describe('"trackStopped" event', () => {
       context('when the MediaStreamTrack emits onended event', () => {
-        it('should emit Track#stopped, passing the instance of Track', () => {
-          track = createTrack(LocalTrack, '1', kind[description]);
+        it('should emit LocalMediaTrack#stopped, passing the instance of LocalMediaTrack', () => {
+          track = crateLocalMediaTrack(LocalMediaTrack, '1', kind[description]);
 
           const stoppedEvent = new Promise((resolve, reject) => {
             track.on('stopped', function(_track) {
@@ -101,7 +101,7 @@ var log = require('../../../../lib/fakelog');
 
     describe('#disable', () => {
       before(() => {
-        track = createTrack(LocalTrack, 'foo', kind[description]);
+        track = crateLocalMediaTrack(LocalMediaTrack, 'foo', kind[description]);
         track.enable = sinon.spy();
         track.disable();
       });
@@ -117,7 +117,7 @@ var log = require('../../../../lib/fakelog');
         var trackEnabledEmitted = false;
 
         before(() => {
-          track =  createTrack(LocalTrack, 'foo', kind[description]);
+          track =  crateLocalMediaTrack(LocalMediaTrack, 'foo', kind[description]);
           track.mediaStreamTrack.enabled = Math.random() > 0.5;
           track.on('disabled', () => trackDisabledEmitted = true);
           track.on('enabled', () => trackEnabledEmitted = true);
@@ -135,7 +135,7 @@ var log = require('../../../../lib/fakelog');
             var eventEmitted = false;
 
             before(() => {
-              track =  createTrack(LocalTrack, 'foo', kind[description]);
+              track =  crateLocalMediaTrack(LocalMediaTrack, 'foo', kind[description]);
               track.mediaStreamTrack.enabled = !enabled;
               track.on(enabled ? 'enabled' : 'disabled', () => eventEmitted = true);
               track.enable(enabled);
@@ -161,7 +161,7 @@ var log = require('../../../../lib/fakelog');
       };
 
       before(function() {
-        track = createTrack(LocalTrack, '1', kind[description]);
+        track = crateLocalMediaTrack(LocalMediaTrack, '1', kind[description]);
         track._createElement = sinon.spy(() => dummyElement);
       });
 
@@ -183,9 +183,9 @@ var log = require('../../../../lib/fakelog');
   });
 });
 
-function createTrack(LocalTrack, id, kind) {
+function crateLocalMediaTrack(LocalMediaTrack, id, kind) {
   var mediaStreamTrack = new MediaStreamTrack(id, kind);
-  return new LocalTrack(mediaStreamTrack, { log: log });
+  return new LocalMediaTrack(mediaStreamTrack, { log: log });
 }
 
 function MediaStreamTrack(id, kind) {
